UPDATE: add in new task to create a package UPDATE: add in new task to convert a list of files to a package structure
The task to create a package takes a version to be used in the package.

The task to convert a list of files to a package structure copies files from a currently retrieved directory, and places them in a new directory with a package.xml file generated, and ready for deployment.